[DRAFT][FEATURE] Add "resolveLinks" option
This feature adds a "resolveLinks" option
to treemenu and listmenu to already build
the full URL within the cached entry, so fluid
does not need to re-generate the link.

The special "_link" property is then filled with
the link.

Only use this feature when you are not really
dealing with special TYPO3 features such as mountpoints.