[TASK] Make page fields to select configurable via configuration for TreeMenu
Add the possibility to provide a list of page fields to select for the TreeMenu
by setting e.g.:
pageFields = uid,pid,title,slug

- If the property is not set, it uses "*" for the select fields.
- If "*" is also in the list, it gets deleted from the list
- empty values wil be removed, so a type like "uid,,pid,title" is covered

The TreeMenu processor is a fast way to use it e.g. for an RESTful API.
In order to create simpler API schemas it´s easier to define which fields are needed from pages.
So the schema does not need to provide all ~100 fields if 5-6 are enough.

The change does not touch that nav_title, hasSubpages and subpages will be populated.